pyproject.toml
[project]
name = "data-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"google-api-python-client>=2.170.0",
	"google-auth>=2.40.0",
    "tqdm>=4.66",
]